Allow fractional # of chunks generated per tick

+++ src/main/java/com/gecgooden/chunkgen/reference/Reference.java
@@ -8,21 +8,21 @@ import org.apache.logging.log4j.Logger;
 import com.gecgooden.chunkgen.util.ChunkPosition;
 
 public class Reference {
-	public static Integer x; 
-	public static Integer z;
-	public static Integer height;
-	public static Integer width;
-	public static Integer numChunksPerTick;
-	
-	public static final String MOD_ID = "chunkgen";
+    public static Integer x;
+    public static Integer z;
+    public static Integer height;
+    public static Integer width;
+    public static double numChunksPerTick;
+
+    public static final String MOD_ID = "chunkgen";
     public static final String VERSION = "1.7.10-1.2.2";
     public static final String GUI_FACTORY = "com.gecgooden.chunkgen.client.gui.GuiFactory";
-    
+
     public static Queue<ChunkPosition> toGenerate;
     public static int startingSize;
     public static int tickDelay;
- 
+
     public static DecimalFormat decimalFormat;
-    
+
     public static Logger logger;
 }
